+++ cg/cli/post_process/post_process.py
@@ -8,7 +8,7 @@ from cg.cli.post_process.utils import get_post_processing_service_from_run_name
 from cg.cli.utils import CLICK_CONTEXT_SETTINGS
 from cg.constants.cli_options import DRY_RUN
 from cg.models.cg_config import CGConfig
-from cg.services.post_processing.abstract_classes import PostProcessingService
+from cg.services.run_devices.abstract_classes import PostProcessingService
 
 LOG = logging.getLogger(__name__)
