+++ cg/services/run_devices/pacbio/data_transfer_service/data_transfer_service.py
@@ -1,25 +1,25 @@
 from pydantic_core._pydantic_core import ValidationError
 
-from cg.services.post_processing.abstract_classes import PostProcessingDataTransferService
-from cg.services.post_processing.error_handler import handle_post_processing_errors
-from cg.services.post_processing.exc import (
+from cg.services.run_devices.abstract_classes import PostProcessingDataTransferService
+from cg.services.run_devices.error_handler import handle_post_processing_errors
+from cg.services.run_devices.exc import (
     PostProcessingRunFileManagerError,
     PostProcessingDataTransferError,
 )
-from cg.services.post_processing.pacbio.data_transfer_service.dto import (
+from cg.services.run_devices.pacbio.data_transfer_service.dto import (
     PacBioDTOs,
     PacBioSampleSequencingMetricsDTO,
     PacBioSequencingRunDTO,
     PacBioSMRTCellDTO,
 )
-from cg.services.post_processing.pacbio.data_transfer_service.utils import (
+from cg.services.run_devices.pacbio.data_transfer_service.utils import (
     get_sample_sequencing_metrics_dtos,
     get_sequencing_run_dto,
     get_smrt_cell_dto,
 )
-from cg.services.post_processing.pacbio.metrics_parser.metrics_parser import PacBioMetricsParser
-from cg.services.post_processing.pacbio.metrics_parser.models import PacBioMetrics
-from cg.services.post_processing.pacbio.run_data_generator.run_data import PacBioRunData
+from cg.services.run_devices.pacbio.metrics_parser.metrics_parser import PacBioMetricsParser
+from cg.services.run_devices.pacbio.metrics_parser.models import PacBioMetrics
+from cg.services.run_devices.pacbio.run_data_generator.run_data import PacBioRunData
 
 
 class PacBioDataTransferService(PostProcessingDataTransferService):
