+++ tests/fixture_plugins/pacbio_fixtures/service_fixtures.py
@@ -3,21 +3,21 @@
 import pytest
 
 from cg.apps.housekeeper.hk import HousekeeperAPI
-from cg.services.post_processing.pacbio.data_storage_service.pacbio_store_service import (
+from cg.services.run_devices.pacbio.data_storage_service.pacbio_store_service import (
     PacBioStoreService,
 )
-from cg.services.post_processing.pacbio.data_transfer_service.data_transfer_service import (
+from cg.services.run_devices.pacbio.data_transfer_service.data_transfer_service import (
     PacBioDataTransferService,
 )
-from cg.services.post_processing.pacbio.housekeeper_service.pacbio_houskeeper_service import (
+from cg.services.run_devices.pacbio.housekeeper_service.pacbio_houskeeper_service import (
     PacBioHousekeeperService,
 )
-from cg.services.post_processing.pacbio.metrics_parser.metrics_parser import PacBioMetricsParser
-from cg.services.post_processing.pacbio.post_processing_service import PacBioPostProcessingService
-from cg.services.post_processing.pacbio.run_data_generator.pacbio_run_data_generator import (
+from cg.services.run_devices.pacbio.metrics_parser.metrics_parser import PacBioMetricsParser
+from cg.services.run_devices.pacbio.post_processing_service import PacBioPostProcessingService
+from cg.services.run_devices.pacbio.run_data_generator.pacbio_run_data_generator import (
     PacBioRunDataGenerator,
 )
-from cg.services.post_processing.pacbio.run_file_manager.run_file_manager import (
+from cg.services.run_devices.pacbio.run_file_manager.run_file_manager import (
     PacBioRunFileManager,
 )
 from cg.store.store import Store
